Update AssociateGoogleEmailData.rb
Added option to override the script's detection of Nuix 7.4 or higher being used.  When this option is enabled, the script will process the data as though you were using a version below 7.4.  This helps with situations where the data was ingested in a version below 7.4, but the case is later migrated and this script is then ran in a version that is 7.4 or higher.